Introduce ScansStack class to track ble rssi trends

+++ Shared/SmartSkating.Dto/Models/BleScanResultDto.cs
@@ -5,7 +5,7 @@ namespace Sanet.SmartSkating.Dto.Models
     public class BleScanResultDto:EntityBase
     {
         public string DeviceAddress { get; set; } = string.Empty;
-        public double Rssi { get; set; }
+        public int Rssi { get; set; }
         public DateTime Time { get; set; }
     }
 }
